+++ src/feature/myBlog/myBlog.tsx
@@ -1,0 +1,12 @@
+import FlexLayoutWithMain from "../../component/block/layout/FlexLayoutWithMain";
+
+export const MyBlog = () => {
+
+    return (
+        <FlexLayoutWithMain>
+            <div>tmp1</div>
+            <div>tmp</div>
+            <div>tmp</div>
+        </FlexLayoutWithMain>
+    );
+};
