+++ src/style/themes.ts
@@ -1,0 +1,25 @@
+import { DefaultTheme } from "styled-components";
+
+const myBlogTheme: DefaultTheme = {
+    color: {
+        bg: {
+            primary: '#F68989',
+            secondary: '#444555',
+        },
+        font: {
+            primary: '',
+            secondary: ''
+        }
+    },
+
+    font: {
+        primary: '',
+        secondary: ''
+    },
+
+    size: {
+
+    }
+};
+
+export { myBlogTheme };
