--- src/styles/themes.ts
@@ -1,25 +1,0 @@
-import { DefaultTheme } from "styled-components";
-
-const myBlogTheme: DefaultTheme = {
-    color: {
-        bg: {
-            primary: '#F68989',
-            secondary: '#F6E7D8'
-        },
-        font: {
-            primary: '',
-            secondary: ''
-        }
-    },
-
-    font: {
-        primary: '',
-        secondary: ''
-    },
-
-    size: {
-
-    }
-};
-
-export { myBlogTheme };
